creates separate index file for FormFooter component
